+++ CHANGELOG.md
@@ -3,11 +3,11 @@
 ## 0.4.9
 
 - Support folder names with URI-invalid characters (not retroactive).
+- Offer to `Export` when removing all bookmarks.
 
 ## 0.4.8
 
 - Add `Search…` command to search for bookmarks.
-- Offer to `Export` when removing all bookmarks.
 
 ## 0.4.7